Create cupcake.code
This program will call three different functions that
change or access the global cupcake menu and cupcake cost variables.
I utilized various conditional statements to error check input statements
and ensure the function ran under specified conditions to improve efficiency
and reduce errors.
Throughout the process of creating this program I encountered challenges when 
trying to calculate the total cost of the userpick cupcakes. 
This lead me to create a an empty list to hold the various costs 
of the chosen cupcakes. 
I learned to understand the difference between local variables and global variables when 
optimizing a defined function with multiple tasks. 
I feel happy with the outcome of this program. 
In conclusion, this program was able to run successfully 
and perform task goals.